requirements.txt
asgiref==3.7.2
Django==5.0.2
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2024.1
ratelimit==2.2.1
sqlparse==0.4.4
